refactored some names to be more uniform.
