+++ app/services/schedule_generator.py
@@ -1,4 +1,4 @@
-from app.models.subjectTime import SubjectTime
+from app.models.subject_time import SubjectTime
 from app.models.week import Week
 from app.models.subject_count import SubjectCount
